Correct the entire screen scrolling issue & add refresh animation on Dashboard screen
I didn't like the fact that the menu button and header title would move with the screen as I scrolled through it so I changed it so that the ScrollView is only on the vital cards.

I also added a refreshing so that the timestamps of the card could be updated in relation to the current time.