Create Hint component & add to Input/TextWithIcon compoent
I created a Hint component so that I can have hints under text inputs.

The reason why I created the component instead of just adding it to the Input is because I will have different types of TextInputs so it's better if the Hint is portable so I have use it in the other TextInputs as well.