Add children prop to the RoundedButton component
Sometimes I want to be able to do other things in a button than show text like showing an icon or something and the simplest way to do that is just to allow the component to take children and render them.

I didn't remove the whole apparentus with the 'text' prop because I would need to go back and modify everywhere the button is used but also move over the styling which is a hassle.