Replace old pimped-out Text with Link component in ValidateVerificationCode container
I created the Link component because I like having links, they are a lighter alternative to buttons and allow me to have pressable things embedded with text.

I had only done it here but since I realized I was going to need to have a link in the Referral screen, I decided to create the component for it.