Implement real submit on AccountProfile screen
Complete the onSubmit method of this screen and make it functional by using the API method to send the data to the server and added the errors prop to the state to keep track of the submission errors.

I also shortened the componentWillMount() because I realized it was uneccesary long... I'm surprised I only caught this now.