Write`onSendCode` in ForgotPassword screen
In order to provide good UX, users need to be able to request for another code to be sent to their number if they feel like it's taking too long.

Since I need to send a code in both ValidateVerificationCode and PhoneNumberVerification container and that both code is pretty much identical, it would not make for good software engineering to re-write the same method twice in different places and that is was smarter for me to "extract" that functionality from those two containers and pass it down from the screen so I can stay DRY.

So I did.