Write `onResendCode` method in ValidateVerificationCode container
I added the functionality by calling the onResendCode prop with the number, so that this is no longer a placeholder and finally works how you would expect it would.

I also mad a few minor UI changes, like the Text not being a Link and thus not being highlighted and increasing the flex size below to 6 because the submit button was too big.