Update PhoneNumberVerification container to use onSendCode prop
I decided to remove all APIs calls from this container and pass that responsability to their parent screens instead because that makes it easier to re-use code.

For example I need to be able to use the same code in both this component and ValidateVerificationCode since I have a button on that container to re-send another code.

While I removed the responsability of the API calls from the containers I also removed the responsability to update the store's state; therefore this container doesn't need to be redux aware anymore.