Update components style prop type from object to array
I was getting a lot of errors because in some cases I was using an array to pass in multiple styles and at first I thought of using OneTypeOf() but then that doesn't work well with default props and also with the value being set.

If an object it can just be set but if an array it needs to be destructured and so the best decision was just to normalize the array everywhere. I prefer to have the flexibility to be able to pass in multiple styles rather than having to combine multiple styles into one object before setting it.